requirements.txt
# Core dependencies
langchain-core
langchain-anthropic
langchain-community
langgraph
langsmith

# UI and Interface
streamlit==1.32.0

# Type hints and annotations
typing-extensions==4.10.0

# API Tools
tavily-python

# Optional but recommended for better performance
python-dotenv